+++ src/main/java/com/example/geektrust/MemberExpense.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class MemberExpense extends Expense {
 
-  public MemberExpense(double amount, Member paidBy, List<SpendFor> spendFor) {
+  public MemberExpense(Integer amount, Member paidBy, List<SpendFor> spendFor) {
     super(amount, paidBy, spendFor);
   }
 }
